backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
anthropic==0.42.0
google-generativeai==0.8.4
python-multipart==0.0.20
cryptography==44.0.0
structlog==24.4.0
Pillow==12.1.1
